Update GUI
Minor, split out menu to own file.
